+++ application/ui/src/features/annotator/annotations/editable-annotation.component.tsx
@@ -7,7 +7,7 @@ import { useZoom } from '../../../components/zoom/zoom';
 import { useSelectedAnnotations } from '../select-annotation-provider.component';
 import { EditBoundingBox } from '../tools/edit-bounding-box/edit-bounding-box.component';
 import { Annotation, Rect } from '../types';
-import { useAnnotation } from './annotation.component';
+import { useAnnotation } from './annotation-context';
 
 interface EditAnnotationProps {
     children: ReactNode;
