+++ application/ui/src/features/annotator/annotations/selectable-annotation.component.tsx
@@ -5,7 +5,7 @@ import { KeyboardEvent, MouseEvent, ReactNode, useEffect, useRef } from 'react';
 
 import { useAnnotator } from '../annotator-provider.component';
 import { useSelectedAnnotations } from '../select-annotation-provider.component';
-import { useAnnotation } from './annotation.component';
+import { useAnnotation } from './annotation-context';
 
 export const SelectableAnnotation = ({ children }: { children: ReactNode }) => {
     const annotation = useAnnotation();
