+++ application/ui/src/features/project/list/new-project-link.component.tsx
@@ -5,7 +5,7 @@ import { Text } from '@geti/ui';
 import { AddCircle } from '@geti/ui/icons';
 import { Link } from 'react-router-dom';
 
-import { paths } from '../../../router';
+import { paths } from '../../../constants/paths';
 
 import classes from './new-project-link.module.scss';
