+++ application/ui/src/routes/error-page/error-page.tsx
@@ -5,7 +5,7 @@ import { Button, Heading, IllustratedMessage, View } from '@geti/ui';
 import { NotFound } from '@geti/ui/icons';
 import { isRouteErrorResponse, useRouteError } from 'react-router';
 
-import { paths } from '../../router';
+import { paths } from '../../constants/paths';
 import { redirectTo } from '../utils';
 
 const useErrorMessage = () => {
