+++ application/ui/src/routes/project/sink.tsx
@@ -27,8 +27,8 @@ import {
     SchemaWebhookSinkConfig,
 } from '../../api/openapi-spec';
 import { RadioDisclosure } from '../../components/radio-disclosure-group/radio-disclosure-group';
+import { paths } from '../../constants/paths';
 import { useProjectIdentifier } from '../../hooks/use-project-identifier.hook';
-import { paths } from '../../router';
 import { ReactComponent as IconFolder } from './../../assets/icons/folder-arrow-right.svg';
 import { ReactComponent as IconMQTT } from './../../assets/icons/mqtt.svg';
 import { ReactComponent as IconRos } from './../../assets/icons/ros.svg';
